Streamline code...
by removing capability to use "rectangular collocation" approach,
where the collocation points are different from the
Gauss-Chebyshev-Lobatto points.  In practice, this is less stable
(reason not entirely clear), so I am always using the usual
collocation points.  In this case, the operators are somewhat simpler
(e.g., the map from function values on the grid to function values at
the collocation points is the identity), and the code now reflects
that.