Compute Jacobian of final state wrt initial state
This capability enables us to use the time domain solver to drive a
Newton iteration where we enforce time periodicity.  The basic idea is
to form a residual that is the difference between the initial
condition and the solution 1 period later (as computed using the time
domain solver).  The Jacobian of this residual can be computed from
the Jacobian added in this commit.

By driving this residual to zero, we find a time periodic solution
that satisfies the (discrete) PDEs.  Given a decent initial guess (as
obtained from a relatively short time domain run or continuation from
a solution with nearby parameters), this iteration should converge
rapidly, making it useful for sensitivity and UQ.