Change state from electron energy (i.e., 1.5*kB*Te) to ne*electron energy...
on Raja's advice.  Qualitatively speaking, the theory is that
ne*energy is smoother and therefore easier to represent.  This may
lead to under-resolution of the temperature, particularly in the
sheath and approaching the sheath edge, where the temperature can be
very large and have sharp features.  However, in this region, the
temperature should not matter very much, since the electron density is
very low.

In practice I observe that using ne*energy is more robust.  Can
converge a case with V0=1kV using 250 Chebyshev modes (which is still
a lot) that I was unable to get to converge with 300 modes when using
the energy as the state variable.