Fix issues related to adding CN time marching, including...
1) The 'strong' version of the electron flux BC was broken by the
residual and jacobian refactoring.  This commit fixes that.  Moreover,
the weak version seems to not work with the shooting method.  This
could be a bug or a formulation problem.  Regardless, for now, only
the strong version is supported by TPS, and running the weak version
in time marching mode will give a warning.  Sorting out the weak
version will have to wait for another day.

2) Fix the stepSensitivity method to work with CN.  Mainly this boils
down to ensuring that the jacobian0 method does its job, which is
complicated by the fact that some data (e.g., the jacobian itself and
the electric potential) are stored in the class.  As a result, making
sure everything is evaluated at a consistent state and time level
takes some hackery (e.g., the ability to solve the poisson equation
for phi in the jacobian routine).  This appears to all be working now
(as evidenced by observed quadratic convergence in the TPS newton
solver).  It could certainly be improved, but... that will have to
wait for another day.