I have chnaged the code to treat BC for ions strongly and not weakly. Unfortunatly, this causes convergence issues with the periodic solver.
This commit changes the default ion flux BC implementation from weak
(where the ion boundary flux is modified by the BC but the PDE is
retained at the boundary point) to strong (where the PDE is discarded
at the boundary point and replaced by the BC equation).  To allow the
original behavior to be recovered, the behavior of the `--weakbc` flag
has been modified.  Previously the `--weakbc` flag changed the
electron flux BC implementation from strong to weak.  However, in
practice, this option was never used.  Thus, it has been changed here
so that `--weakbc` now modifies the ion flux BC.  This enables the
original behavior to be reproduced even though the default has
changed.